Added MODEL_ALIAS to config files and in scripts
This option makes it easier to manage different implementations of the same model.
For example, there may be small differences between two versions of model code,
which have nothing to do with the model physics, but simply with some run-time
options (such as REWIND on or off in VIC). This allows the user to specify the
same MODEL_ALIAS for these two executables, which means that the directory
structures and file naming will still be identical. For now this has not be
carried forward in naming the logs, so that the logs will still have separate
names.

Be aware that if you give all models the same alias, they are likely to overwrite
each other's output, so use this option wisely.